# Frontend/RepoGeist/openapi.json
{
    "openapi":  "3.1.1",
    "info":  {
                 "title":  "RepoGeist | v1",
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5221/"
                    }
                ],
    "paths":  {
...
                  "/api/repos":  {
                                     "get":  {
                                                 "tags":  [
                                                              "Repos"
                                                          ],
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "OK",
                                                                               "content":  {
                                                                                               "text/plain":  {
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "array",
                                                                                                                                 "items":  {
                                                                                                                                               "$ref":  "#/components/schemas/RepoCardDto"
                                                                                                                                           }
                                                                                                                             }
                                                                                                              },
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "type":  "array",
                                                                                                                                       "items":  {
                                                                                                                                                     "$ref":  "#/components/schemas/RepoCardDto"
                                                                                                                                                 }
                                                                                                                                   }
                                                                                                                    },
                                                                                               "text/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "type":  "array",
                                                                                                                                "items":  {
                                                                                                                                              "$ref":  "#/components/schemas/RepoCardDto"
                                                                                                                                          }
                                                                                                                            }
                                                                                                             }
                                                                                           }
                                                                           }
                                                               }
                                             },
                                     "post":  {
                                                  "tags":  [
                                                               "Repos"
                                                           ],
                                                  "requestBody":  {
                                                                      "content":  {
                                                                                      "application/json":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/CreateRepoRequestDto"
                                                                                                                          }
                                                                                                           },
                                                                                      "text/json":  {
                                                                                                        "schema":  {
                                                                                                                       "$ref":  "#/components/schemas/CreateRepoRequestDto"
                                                                                                                   }
                                                                                                    },
                                                                                      "application/*+json":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/CreateRepoRequestDto"
                                                                                                                            }
                                                                                                             }
                                                                                  },
                                                                      "required":  true
                                                                  },
                                                  "responses":  {
                                                                    "200":  {
                                                                                "description":  "OK",
                                                                                "content":  {
                                                                                                "text/plain":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/RepoCardDto"
                                                                                                                              }
                                                                                                               },
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/RepoCardDto"
                                                                                                                                    }
                                                                                                                     },
                                                                                                "text/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/RepoCardDto"
                                                                                                                             }
                                                                                                              }
                                                                                            }
                                                                            }
                                                                }
                                              }
                                 },
                  "/api/repos/{owner}/{name}":  {
                                                    "get":  {
                                                                "tags":  [
                                                                             "Repos"
                                                                         ],
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "owner",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   },
                                                                                   {
                                                                                       "name":  "name",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  {
                                                                                                              "text/plain":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                            }
                                                                                                                             },
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                                  }
                                                                                                                                   },
                                                                                                              "text/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                           }
                                                                                                                            }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                                },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateRepoRequestDto":  {
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "repoUrl":  {
                                                                                                       "type":  "string"
                                                                                                   }
                                                                                   }
                                                                },
...
                                       "HealthStatus":  {
                                                            "type":  "integer"
                                                        },
...
                                       "RepoCardDto":  {
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "id":  {
                                                                                         "type":  "string",
                                                                                         "format":  "uuid"
                                                                                     },
                                                                              "owner":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                              "name":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                              "description":  {
                                                                                                  "type":  [
                                                                                                               "null",
                                                                                                               "string"
                                                                                                           ]
                                                                                              },
                                                                              "topics":  {
                                                                                             "type":  "array",
                                                                                             "items":  {
                                                                                                           "type":  "string"
                                                                                                       }
                                                                                         },
                                                                              "primaryLanguage":  {
                                                                                                      "type":  [
                                                                                                                   "null",
                                                                                                                   "string"
                                                                                                               ]
                                                                                                  },
                                                                              "liveDemoUrl":  {
                                                                                                  "type":  [
                                                                                                               "null",
                                                                                                               "string"
                                                                                                           ]
                                                                                              },
                                                                              "healthStatus":  {
                                                                                                   "$ref":  "#/components/schemas/HealthStatus"
                                                                                               }
                                                                          }
                                                       },
...
                                   }
                   },
...
}